1)Added reusable step in feature file for clicking  on link TV, Appliances, Electronics
2)Added reusable Wait method to avoid immediate termination of browser.

+++ src/main/java/com/selenium/cbh/Utilities.java
@@ -110,8 +110,8 @@ public class Utilities extends base{
     }
 
 
-    public static int minutesToMilliSecond(int min) {
-        int sec;
+    public static long minutesToMilliSecond(long min) {
+        long sec;
 
         sec = min * 60 * 1000;
